fix: handle the type of box's additional_details in a different way
