+++ Pipfile
@@ -16,7 +16,6 @@ pylint = "~=2.15.0"
 [packages]
 policyuniverse = "==1.4.0.20220110"
 requests = "~=2.27"
-panther-core = "*"
 
 
 [requires]
